refactor: organize code into domains
Organizing by application layers no longer works well because the code
is split across those layers unevenly and it leads only to code
fragmentation. Organizing into domains works much better since the
global space is not polluted.
