+++ src/domains/conjugations/conjugation.ts
