+++ src/domains/vocabulary/Results.tsx
