+++ src/domains/vocabulary/Tester.tsx
@@ -1,8 +1,8 @@
 import { Show, createEffect, createSignal, onCleanup } from 'solid-js';
+import { Progress } from '../../components/Progress';
 import { WordTranslation } from '../../parser/simple-md-parser';
 import { mergeWords } from '../../util/merge-arrays';
 import { nextWord } from '../../worder/worder';
-import { Progress } from '../Progress';
 import { WriteTester } from './WriteTester';
 
 export type TestMode = 'guess' | 'write';
