+++ src/domains/vocabulary/WordsInput.module.css
